samples/backend/python/quart/requirements.txt
openai
quart[dotenv]
azure.identity